# chapter13/whatstore/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Whatstore API",
        "version": "0.1"
    },
    "paths": {
...
        "/inventoryapi/productapi/{key}": {
            "get": {
                "operationId": "c5fa61b89c9f23e9fc68cda3ce66a93a",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "product code",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get a product"
                    }
                }
            },
            "put": {
                "operationId": "14e0a94eae01d5580d2fe53c6cdf12fd",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "product code",
                        "required": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "code_discount"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "float"
                                    },
                                    "code_discount": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "update a product"
                    }
                }
            },
            "post": {
                "operationId": "f49d93870a886459c50f6b1f53e3edd1",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "product code",
                        "required": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "code_discount"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "float"
                                    },
                                    "code_discount": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "insert a product"
                    }
                }
            },
            "delete": {
                "operationId": "c705565d7d60524e8f8cfd41a4ea7b78",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "product code",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "remove a product"
                    }
                }
            }
        },
...
        "/inventoryapi/roleapi/{key}": {
            "get": {
                "operationId": "abebeb460ce9c0998ae93ea5c1f8db7d",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "role code",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get a role"
                    }
                }
            },
            "put": {
                "operationId": "7eda769a39b09927d8e5a4812c719cf2",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "role code",
                        "required": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "nickname",
                                    "password"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "nickname": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "update a role"
                    }
                }
            },
            "post": {
                "operationId": "69b19b13eee9a41ea82696b7abbe6293",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "role code",
                        "required": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "nickname",
                                    "password"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "nickname": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "insert a role"
                    }
                }
            },
            "delete": {
                "operationId": "83ddf6080dd2ba6788e958ff6c1e8723",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "role code",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "remove a role"
                    }
                }
            }
        },
...
        "/inventoryapi/roleresourceapi/{key}": {
            "post": {
                "operationId": "5bda5c9871b8a1eac69d4e360155817d",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "role resource code",
                        "required": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "code_role",
                                    "code_resource"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "integer"
                                    },
                                    "code_role": {
                                        "type": "integer"
                                    },
                                    "code_resource": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "insert a role resource"
                    }
                }
            },
            "delete": {
                "operationId": "e27a89e3cd61c7085799718ac8ba8e67",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "role resource code",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "remove a role resource"
                    }
                }
            }
        },
...
    }
}